+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.4.0
 - Better String subscription methods.
+- More URL utils.
 
 ### 1.3.1
 - Access level fixes.