+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### 1.4.0
 - Better String subscription methods.
 - More URL utils.
+- More Character utils.
+- More ProcessInfo utils.
+- Bugfixes.
 
 ### 1.3.1
 - Access level fixes.