+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-### 1.X.Y (Work in progress)
+### 1.5.0 (Work in progress)
 - Added Font utils for bold and italic traits.
 - Added shuffling methods for collections and sequences.
+- Better random string generation methods.
+- Bugfixes.
 
 ### 1.4.0
 - Better String subscription methods.